backend/app/core/config.py
from pydantic_settings import BaseSettings
from typing import Optional, List
import os
from pathlib import Path
from dotenv import load_dotenv

ROOT_DIR = Path(__file__).resolve().parents[2]
load_dotenv(ROOT_DIR / ".env")

class Settings(BaseSettings):
    PROJECT_NAME: str = "TrustSeal IoT"
    API_V1_STR: str = "/api/v1"
    SECRET_KEY: str = os.getenv("SECRET_KEY", "your-secret-key-here-change-in-production")
    ALGORITHM: str = "HS256"
    ACCESS_TOKEN_EXPIRE_MINUTES: int = 60 * 24 * 7  # 7 days
    VERIFICATION_TOKEN_EXPIRE_MINUTES: int = int(os.getenv("VERIFICATION_TOKEN_EXPIRE_MINUTES", "60"))
    BACKEND_CORS_ORIGINS: str = os.getenv(
        "BACKEND_CORS_ORIGINS",
        "http://localhost:5173,http://127.0.0.1:5173,https://trust-seal-tawny.vercel.app",
    )
    BACKEND_CORS_ORIGIN_REGEX: Optional[str] = os.getenv("BACKEND_CORS_ORIGIN_REGEX")
    DATABASE_URL_OVERRIDE: Optional[str] = os.getenv("DATABASE_URL")
    
    # Database
    POSTGRES_SERVER: str = os.getenv("POSTGRES_SERVER", "localhost")
    POSTGRES_USER: str = os.getenv("POSTGRES_USER", "postgres")
    POSTGRES_PASSWORD: str = os.getenv("POSTGRES_PASSWORD", "postgres")
    POSTGRES_DB: str = os.getenv("POSTGRES_DB", "trustseal")
    POSTGRES_PORT: Optional[int] = os.getenv("POSTGRES_PORT", 5432)
    POSTGRES_SSLMODE: Optional[str] = os.getenv("POSTGRES_SSLMODE", "prefer")
    WS_REQUIRE_AUTH: bool = os.getenv("WS_REQUIRE_AUTH", "false").lower() == "true"
    REALTIME_QUEUE_MAXSIZE: int = int(os.getenv("REALTIME_QUEUE_MAXSIZE", "5000"))

    OPENROUTER_API_KEY: Optional[str] = os.getenv("OPENROUTER_API_KEY")
    OPENROUTER_BASE_URL: str = os.getenv("OPENROUTER_BASE_URL", "https://openrouter.ai/api/v1")
    OPENROUTER_MODEL: str = os.getenv("OPENROUTER_MODEL", "anthropic/claude-3.5-haiku")
    OPENROUTER_TIMEOUT_SECONDS: int = int(os.getenv("OPENROUTER_TIMEOUT_SECONDS", "30"))
    OPENROUTER_MAX_TOKENS: int = int(os.getenv("OPENROUTER_MAX_TOKENS", "512"))
    OPENROUTER_SITE_URL: Optional[str] = os.getenv("OPENROUTER_SITE_URL")
    OPENROUTER_APP_NAME: str = os.getenv("OPENROUTER_APP_NAME", "TrustSeal IoT")
    OPENAI_API_KEY: Optional[str] = os.getenv("OPENAI_API_KEY")
    OPENAI_BASE_URL: Optional[str] = os.getenv("OPENAI_BASE_URL")
    RAG_COLLECTION_NAME: str = os.getenv("RAG_COLLECTION_NAME", "trustseal_ops")
    RAG_EMBEDDING_MODEL: str = os.getenv("RAG_EMBEDDING_MODEL", "openai/text-embedding-3-small")
    RAG_EMBEDDING_DIMENSION: int = int(os.getenv("RAG_EMBEDDING_DIMENSION", "1536"))
    RAG_TOP_K: int = int(os.getenv("RAG_TOP_K", "6"))
    RAG_MIN_RELEVANCE: float = float(os.getenv("RAG_MIN_RELEVANCE", "0.25"))
    RAG_TEMPERATURE: float = float(os.getenv("RAG_TEMPERATURE", "0.2"))
    RAG_MAX_SHIPMENTS: int = int(os.getenv("RAG_MAX_SHIPMENTS", "300"))
    RAG_MAX_DEVICES: int = int(os.getenv("RAG_MAX_DEVICES", "300"))
    RAG_MAX_SENSOR_LOGS: int = int(os.getenv("RAG_MAX_SENSOR_LOGS", "500"))
    RAG_MAX_CUSTODY_CHECKPOINTS: int = int(os.getenv("RAG_MAX_CUSTODY_CHECKPOINTS", "500"))
    RAG_TOOL_MAX_STEPS: int = int(os.getenv("RAG_TOOL_MAX_STEPS", "5"))
    CHAT_MEMORY_MAX_TURNS: int = int(os.getenv("CHAT_MEMORY_MAX_TURNS", "8"))
    CHAT_MEMORY_TTL_MINUTES: int = int(os.getenv("CHAT_MEMORY_TTL_MINUTES", "240"))
    TEMPERATURE_THRESHOLD_C: float = float(os.getenv("TEMPERATURE_THRESHOLD_C", "8"))
    AGENTIC_VECTOR_COLLECTION: str = os.getenv("AGENTIC_VECTOR_COLLECTION", "trustseal_agentic_docs")
    AGENTIC_TOP_K: int = int(os.getenv("AGENTIC_TOP_K", "5"))
    AGENTIC_SIMILARITY_THRESHOLD: float = float(os.getenv("AGENTIC_SIMILARITY_THRESHOLD", "0.35"))
    AGENTIC_MMR_FETCH_K: int = int(os.getenv("AGENTIC_MMR_FETCH_K", "20"))
    AGENTIC_MMR_LAMBDA: float = float(os.getenv("AGENTIC_MMR_LAMBDA", "0.5"))
    AGENTIC_LLM_MODEL: str = os.getenv("AGENTIC_LLM_MODEL", os.getenv("OPENROUTER_MODEL", "openai/gpt-4o-mini"))
    AGENTIC_EMBEDDING_MODEL: str = os.getenv(
        "AGENTIC_EMBEDDING_MODEL",
        os.getenv("RAG_EMBEDDING_MODEL", "openai/text-embedding-3-small"),
    )
    AGENTIC_CHUNK_SIZE: int = int(os.getenv("AGENTIC_CHUNK_SIZE", "800"))
    AGENTIC_CHUNK_OVERLAP: int = int(os.getenv("AGENTIC_CHUNK_OVERLAP", "120"))
    AGENTIC_BATCH_SIZE: int = int(os.getenv("AGENTIC_BATCH_SIZE", "32"))
    AGENTIC_MAX_RESPONSE_TOKENS: int = int(os.getenv("AGENTIC_MAX_RESPONSE_TOKENS", "450"))
    AGENTIC_TEMPERATURE: float = float(os.getenv("AGENTIC_TEMPERATURE", "0.2"))
    AGENTIC_MAX_TOOL_STEPS: int = int(os.getenv("AGENTIC_MAX_TOOL_STEPS", "4"))
    AGENTIC_POOL_MIN_SIZE: int = int(os.getenv("AGENTIC_POOL_MIN_SIZE", "2"))
    AGENTIC_POOL_MAX_SIZE: int = int(os.getenv("AGENTIC_POOL_MAX_SIZE", "10"))
    AGENTIC_SHORT_MEMORY_WINDOW: int = int(os.getenv("AGENTIC_SHORT_MEMORY_WINDOW", "6"))
    AGENTIC_SHORT_MEMORY_TTL_MINUTES: int = int(os.getenv("AGENTIC_SHORT_MEMORY_TTL_MINUTES", "240"))
    AGENTIC_LONG_MEMORY_TOP_K: int = int(os.getenv("AGENTIC_LONG_MEMORY_TOP_K", "4"))
    
    @property
    def DATABASE_URL(self) -> str:
        if self.DATABASE_URL_OVERRIDE:
            return self.DATABASE_URL_OVERRIDE
        if self.POSTGRES_SSLMODE == "require":
            return f"postgresql://{self.POSTGRES_USER}:{self.POSTGRES_PASSWORD}@{self.POSTGRES_SERVER}:{self.POSTGRES_PORT}/{self.POSTGRES_DB}?sslmode=require"
        return f"postgresql://{self.POSTGRES_USER}:{self.POSTGRES_PASSWORD}@{self.POSTGRES_SERVER}:{self.POSTGRES_PORT}/{self.POSTGRES_DB}"

    @property
    def CORS_ORIGINS(self) -> List[str]:
        # Browser Origin header never has a trailing slash. Normalize to avoid false CORS mismatches.
        raw_origins = [origin.strip().rstrip("/") for origin in self.BACKEND_CORS_ORIGINS.split(",")]
        return [origin for origin in raw_origins if origin]
    
    class Config:
        case_sensitive = True

settings = Settings()
